refactor(cli, agent): extract shared council suggest logic and restructure cli module
Phase 0: Extract shared suggest_council() into gglib-agent::council::suggest
- Move suggest orchestration, parse_suggested_council(), strip_markdown_json()
  from CLI and Axum into a single shared module
- Add SuggestedCouncil::into_config() conversion method
- Both CLI and Axum now call the shared function (DRY)

Phase 1: Split CLI council handler into modular file tree
- council.rs → council/mod.rs (orchestration + routing)
- council/stream.rs (live ANSI renderer + temperature_fg)
- council/render.rs, repl.rs, editor.rs (stubs for interactive REPL)
- Extract run_council_config() helper for future REPL reuse
- All files under 150 LOC
