+++ crates/gglib-agent/src/council/mod.rs
@@ -16,6 +16,7 @@
 //! | `history.rs`      | Per-turn context builder (identity + transcript)    |
 //! | `stream_bridge.rs`| `AgentEvent` → `CouncilEvent` mapper                |
 //! | `orchestrator.rs` | Round×agent loop driver + synthesis dispatch         |
+//! | `suggest.rs`      | `suggest_council()` — shared suggest orchestration  |
 
 pub mod config;
 pub mod events;
@@ -24,6 +25,7 @@ pub mod orchestrator;
 pub mod prompts;
 pub mod state;
 pub mod stream_bridge;
+pub mod suggest;
 
 pub use config::{CouncilAgent, CouncilConfig, SuggestedCouncil};
 pub use events::{COUNCIL_EVENT_CHANNEL_CAPACITY, CouncilEvent};
@@ -31,3 +33,4 @@ pub use orchestrator::run as run_council;
 pub use prompts::{contentiousness_tier_label, contentiousness_to_instruction};
 pub use state::{AgentContribution, CouncilState, extract_core_claim};
 pub use stream_bridge::{bridge_agent_events, emit_turn_complete};
+pub use suggest::suggest_council;
