+++ crates/gglib-cli/src/handlers/council/editor.rs
@@ -1,0 +1,1 @@
+//! Agent state mutation logic for the council editing REPL.
