+++ crates/gglib-cli/src/handlers/council/render.rs
@@ -1,0 +1,1 @@
+//! Static ANSI rendering for council summary and agent cards.
