+++ crates/gglib-cli/src/handlers/council/repl.rs
@@ -1,0 +1,1 @@
+//! Rustyline-based interactive editing REPL for council configuration.
